Fix overlay context handling; Return error when overlay path is used a second time
